Update to network_learning.py
Added function for BCM rule, as well as modifications to the main and learn functions to fit the change. Essentially three new parameters, theta, tau_t, and rule, which are initialized in main and passed through to learn. The rule variable just allows for a static initial rule choice, and theta and tau_t are necessary for the BCM step (and are ultimately passed to that function if BCM is the selected rule).