Class for global rule objects
Made this file to keep the main program a little cleaner. I'll modify the main program as well, but to use this, import it by name, and then outside of the training loop create an object with desired arguments, and inside the loop call the run() function with parameters f, W, and step.